+++ lib/index.js
@@ -1,5 +1,5 @@
 const Pack = require('../package.json');
-const PluginRegistrationV2 = require('./pluginRegistration.js');
+const PluginRegistrationV2 = require('./v2/pluginRegistration.js');
 
 
 exports.plugin = {
