+++ lib/v2/builder.js
@@ -1,14 +1,14 @@
 const Hoek = require('@hapi/hoek');
 const Joi = require('joi');
 const JSONDeRef = require('json-schema-ref-parser');
-const Filter = require('../lib/filter');
-const Group = require('../lib/group');
-const Sort = require('../lib/sort');
-const Info = require('../lib/info');
-const Paths = require('../lib/paths');
-const Tags = require('../lib/tags');
-const Validate = require('../lib/validate');
-const Utilities = require('../lib/utilities');
+const Filter = require('./filter');
+const Group = require('./group');
+const Sort = require('./sort');
+const Info = require('./info');
+const Paths = require('./paths');
+const Tags = require('./tags');
+const Validate = require('./validate');
+const Utilities = require('./utilities');
 
 const builder = (module.exports = {});
 const internals = {};
