+++ lib/v2/defaults.js
