+++ lib/v2/definitions.js
@@ -1,5 +1,5 @@
 const Hoek = require('@hapi/hoek');
-const Utilities = require('../lib/utilities');
+const Utilities = require('./utilities');
 
 const internals = {};
