+++ lib/v2/filter.js
