+++ lib/v2/group.js
@@ -1,4 +1,4 @@
-const Utilities = require('../lib/utilities');
+const Utilities = require('./utilities');
 
 const group = (module.exports = {});
