+++ lib/v2/info.js
