+++ lib/v2/parameters.js
@@ -1,5 +1,5 @@
 const Hoek = require('@hapi/hoek');
-const Utilities = require('../lib/utilities');
+const Utilities = require('./utilities');
 
 const parameters = (module.exports = {});
