+++ lib/v2/paths.js
@@ -1,11 +1,11 @@
 const Hoek = require('@hapi/hoek');
 const Joi = require('joi');
 
-const Parameters = require('../lib/parameters');
-const Definitions = require('../lib/definitions');
-const Properties = require('../lib/properties');
-const Responses = require('../lib/responses');
-const Utilities = require('../lib/utilities');
+const Parameters = require('./parameters');
+const Definitions = require('./definitions');
+const Properties = require('./properties');
+const Responses = require('./responses');
+const Utilities = require('./utilities');
 
 const internals = {};
