+++ lib/v2/pluginRegistration.js
@@ -4,9 +4,9 @@ const Path = require('path');
 const { join, sep } = require('path');
 const swaggerUiAssetPath = require('swagger-ui-dist').getAbsoluteFSPath();
 
-const Defaults = require('../lib/defaults');
-const Builder = require('../lib/builder');
-const Utilities = require('../lib/utilities');
+const Defaults = require('./defaults');
+const Builder = require('./builder');
+const Utilities = require('./utilities');
 
 // schema for plug-in properties
 const schema = Joi.object({
@@ -90,7 +90,7 @@ module.exports = (server, options) => {
     // It should be overwritten only if explicitly passed to the plugin options.
     const validateOption = { validate: options.validate }
     const settings = Hoek.applyToDefaults(Defaults, options, { nullOverride: true });
-    const publicDirPath = Path.resolve(__dirname, '..', 'public');
+    const publicDirPath = Path.resolve(__dirname, '../..', 'public');
 
     // avoid breaking behaviour with previous version
     if (!options.routesBasePath && options.swaggerUIPath) {
