+++ lib/v2/properties.js
@@ -1,6 +1,6 @@
 const Hoek = require('@hapi/hoek');
-const Definitions = require('../lib/definitions');
-const Utilities = require('../lib/utilities');
+const Definitions = require('./definitions');
+const Utilities = require('./utilities');
 
 const internals = {};
