+++ lib/v2/responses.js
@@ -1,9 +1,9 @@
 const HTTPStatus = require('http-status');
 const Hoek = require('@hapi/hoek');
-const Definitions = require('../lib/definitions');
-const Properties = require('../lib/properties');
+const Definitions = require('./definitions');
+const Properties = require('./properties');
 
-const Utilities = require('../lib/utilities');
+const Utilities = require('./utilities');
 
 const internals = {};
