+++ lib/v2/sort.js
@@ -1,4 +1,4 @@
-const Utilities = require('../lib/utilities');
+const Utilities = require('./utilities');
 
 const sort = (module.exports = {});
