+++ lib/v2/validate.js
