+++ test/Integration/cache-test.js
@@ -3,7 +3,7 @@ const Joi = require('joi');
 const Lab = require('@hapi/lab');
 
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
