+++ test/Integration/dereference-test.js
@@ -1,9 +1,9 @@
 const Code = require('@hapi/code');
 const Joi = require('joi');
 const Lab = require('@hapi/lab');
-const Builder = require('../../lib/builder.js');
+const Builder = require('../../lib/v2/builder.js');
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
