+++ test/Integration/file-test.js
@@ -2,7 +2,7 @@ const Code = require('@hapi/code');
 const Joi = require('joi');
 const Lab = require('@hapi/lab');
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
