+++ test/Integration/http2-test.js
@@ -1,7 +1,7 @@
 const Code = require('@hapi/code');
 const Lab = require('@hapi/lab');
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 const Http2 = require('http2');
 const Fs = require('fs').promises;
 const Path = require('path');
