+++ test/Integration/info-test.js
@@ -1,7 +1,7 @@
 const Code = require('@hapi/code');
 const Lab = require('@hapi/lab');
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
