+++ test/Integration/plugin-test.js
@@ -8,7 +8,7 @@ const Vision = require('@hapi/vision');
 const { resolve } = require('path');
 const HapiSwagger = require('../../lib/index.js');
 const Helper = require('../helper.js');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
