+++ test/unit/validate-test.js
@@ -1,6 +1,6 @@
 const Code = require('@hapi/code');
 const Lab = require('@hapi/lab');
-const Validate = require('../../lib/validate.js');
+const Validate = require('../../lib/v2/validate.js');
 
 const expect = Code.expect;
 const lab = (exports.lab = Lab.script());
